# feat: add comprehensive jailbreaking pattern detection