@mickdarling 2025-07-17
## ✅ All Review Recommendations Addressed

I've implemented all three recommendations from the review:

### 1. ✅ Expanded hypothetical_bypass lookbehind (20 → 40 chars)
- Pattern now catches longer phrases between "hypothetical" and restriction keywords
- Example: "hypothetically speaking, in a scenario where you have no restrictions"

### 2. ✅ Fixed skipped test 
- Root cause: Was using wrong test runner configuration
- Solution: Run with `npm run test:security` instead of `npm test`
- All 12 jailbreaking tests now passing

### 3. ✅ Added comprehensive performance benchmarks
New test file: `test/security/performance-benchmark.test.ts` with:
- Content size scaling tests (1KB to 500KB)
- Pattern count impact analysis
- Line tracking performance
- Memory usage considerations

**Performance Results:**
- **API requests (3KB)**: <0.1ms 🚀
- **Large documents (50KB)**: <1ms 🚀
- **Very large content (500KB)**: ~5ms
- **Average per pattern**: ~0.02ms
- **Scaling**: Linear with content size

The scanner shows excellent performance characteristics suitable for production use.

### Test Summary
All 136 tests passing:
- Unit Tests: 42 ✅
- Integration Tests: 32 ✅  
- Security Tests: 62 ✅ (including 15 new performance tests)

This PR is ready for final review and merge\!